+++ internal/config/config.go
@@ -30,6 +30,7 @@ type Options struct {
 	AFKThresholdSeconds int         `json:"afk_threshold_seconds,omitempty"`
 	DefaultVolume       int         `json:"default_volume,omitempty"`
 	Log                 bool        `json:"log,omitempty"`
+	Echo                bool        `json:"echo,omitempty"`
 	Cooldown            bool        `json:"cooldown,omitempty"`
 	CooldownSeconds     int         `json:"cooldown_seconds,omitempty"`
 	Credentials         Credentials `json:"credentials,omitempty"`